fix: separando as interfaces, centralizando na pasta ultis/types
