+++ src/app/gerenciamento/turma/page.tsx
@@ -2,7 +2,8 @@
 
 import React, { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
-import Lista, { Usuario } from "@/components/lista";
+import { Usuario } from "@/utils/types";
+import Lista from "@/components/lista";
 import { useRouter } from "next/navigation";
 import { toast } from "sonner";
 import TextField from "@/components/input/textField";
