+++ src/components/lista.tsx
@@ -20,17 +20,10 @@ import ActionModal from "./modal/actionModal";
 
 import { Textarea } from "./ui/textarea";
 import { toast } from "sonner";
-import { Conselho } from "@/app/conselho/pagina";
-
+import { Conselho } from "@/utils/types";
+import { Usuario } from "@/utils/types";
 import { Label } from "./ui/label";
 import { useLocalStorage } from "usehooks-ts";
-
-export interface Usuario {
-  id: number;
-  nome: string;
-  email: string;
-  role: string;
-}
 
 interface ListaProps {
   className?: string;
